u2fd: add more logging to prepare dogfooding
Dump in the logs the mode and the certificate (if the verbose logging is
enabled) in order to be able to check the dogfooders setup in the
feedback report logs.

BUG=b:35545754
TEST=manual, enable verbose logging by using 'u2f_flags g2f verbose'
command in crosh, and stare at the u2fd log in /var/log/messages,
the certificate content is properly dumped.

Change-Id: I64c2434e4e796c24d68e522872c5137e1ea440a2
Reviewed-on: https://chromium-review.googlesource.com/687914
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Trybot-Ready: Vincent Palatin <vpalatin@chromium.org>