chromeos-config: Add alternative whitelabel support
With this schema addition, we can have whitelabels which are not models,
but are all in the same model, with just the keys (and perhaps a few other
things) being different. This is useful for situations where we have (say)
90 different whitelabels and the overhead of seeing these all in GoldenEye
is too great.

BUG=chromium:781928
TEST=PYTHONPATH=~/c python libcros_config_host_unittest.py
Manually tested on robo
Also manually tested by making robo a whitelabel and setting up vpd
Compared reset of the new and old schema and see that the firmware update
produced is identical, except for crbug.com/784566

Change-Id: Iced75f4257db5b0f9a7255959f468b722cd8d919
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/763907
Reviewed-by: Jason Clinton <jclinton@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/794710
Commit-Queue: Marco Chen <marcochen@chromium.org>
Tested-by: Marco Chen <marcochen@chromium.org>