chromeos-config: Form the ebuild name from the overlay name
At present we use the 'base model' to form the chromeos-firmware-xxx name
where xxx is the base model. This name is part of the URI whence the
firmware is collected. The 'base model' is the model that actually
provides the firmware, taking account of firmware sharing.

In practice we don't always store the firmware in this place, however. A
better way to determine the correct name is to extract it from the
'overlay-xxx-private' overlay name. This works for coral and reef. It
should also work for reef migration, but that remains to be seen.

BUG=b:72337771
TEST=./run_tests.sh

Change-Id: Ia0a80eafb611906d15930ccf8a6480d64ebcc8b8
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/886828
Tested-by: Patrick Berny <pberny@chromium.org>
Reviewed-by: Patrick Berny <pberny@chromium.org>
(cherry picked from commit 9f320596ef131c76a1bbed0b77d6f49078f64eb7)
Reviewed-on: https://chromium-review.googlesource.com/900385
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>
Commit-Queue: Patrick Berny <pberny@chromium.org>
Trybot-Ready: Patrick Berny <pberny@chromium.org>