power: Separate WiFi power settings from tablet mode
This CL adds a WifiTransmitPower enum that is used to change the power
of the WiFi antenna, as a separate concept from TabletMode. In this CL,
WifiTransmitPower is derived directly from TabletMode, but this is preparatory
work to enable different information sources (namely, proximity sensor data)
to be used as a source of transmit power.

TEST=run powerd unit tests
BUG=b:74586744

Change-Id: I46b296af0a65e7595e968bcee6acc077d9ea21cd
Signed-off-by: Enrico Granata <egranata@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1148681
Reviewed-by: Dan Erat <derat@chromium.org>
(cherry picked from commit fe1410019b86ec7b8451d8dff49974746ed8c8c3)
Reviewed-on: https://chromium-review.googlesource.com/1170103