biod: log template/hardware version mismatch
When loading a record, read the version of the template and ensure that
it is compatible with the version supported by the hardware. Log an
error if there is a mismatch.
Do not throw an error so that the template is not removed from the
user's view despite still being present on the device.

BUG=b:116396493
BUG=b:116979455
TEST=see "Format mismatch between template (2) and hardware (1)" in
biod's log

Change-Id: I4601ddb0fe3a8491d5b767dd4ada962bfb5fa301
Reviewed-on: https://chromium-review.googlesource.com/1253192
Commit-Ready: Nicolas Norvez <norvez@chromium.org>
Tested-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>
(cherry picked from commit 4929d6da39f4ff2ae403883f6cfbeeb0423d0224)
Reviewed-on: https://chromium-review.googlesource.com/c/1362335
Reviewed-by: YH Lin <yueherngl@chromium.org>
Commit-Queue: YH Lin <yueherngl@chromium.org>
Tested-by: YH Lin <yueherngl@chromium.org>