biod: Add bio_wash binary to reset entropy
This CL introduces a bio_wash program, which clears & re-initializes
any internal crypto entropy/secret of the biometric managers which might
be used for encrypting data handled by the corresponding sensor.

Currently, only the CrosFpBiometricsManager's ResetEntropy() function is
called, but additional BiometricsManager instances can be added.

The bio_wash command is intended to be invoked only during powerwash and
other scenarios where we clobber the stateful partition.

NOTE: Since this CL is a cherry-pick from master, the difference b/w the
two CLs is the absence of a BUILD.gn file (not present in the release
branch). Therefore this CL was modified to add the build target to the
biod.gyp file instead. There is not other change to the source code.

BUG=b:117908009
TEST=1. Build and execute manually.
     2. Make addition to run script during power wash. Check that
     powerwash isn't broken.

Change-Id: Ia8537e85a8fa1a338d2ac75e8146cb1330135c6b
Reviewed-on: https://chromium-review.googlesource.com/1290298
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Prashant Malani <pmalani@google.com>
Reviewed-by: Prashant Malani <pmalani@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/1310540
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Tested-by: Prashant Malani <pmalani@chromium.org>
Commit-Queue: Prashant Malani <pmalani@chromium.org>