metrics: memd: tolerate variations in /proc/vmstat
The old code did not account correctly for different
names of memory zones depending on architecture and
kernel version.  This CL protects against such variations.
Also accumulates the page allocation numbers across
zones, which is usually how we look at these numbers.

This also adds a test for reading /proc/vmstat, and,
as a side effect, cleans up the test environment setup
and teardown.  The main change is the use of a unique
temp directory for the testing root, which allows all
tests to run in parallel.

BUG=chromium:896842
TEST=new unit test, plus ran on minnie and cyan

Change-Id: I4ad0916485df8ad11aa9176817ca48968b929392
Reviewed-on: https://chromium-review.googlesource.com/c/1294812
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>
Commit-Queue: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>