metrics: memd: R70 only: exit(0) on panic to avoid crashes
Because we aren't fully confident that there isn't some other
bug which could generate a lot of crashes, and because the
functionality of memd is limited to data gathering and thus
not critical to the operation, we exit with status 0 on panic
therefore avoiding a crash report from the anomaly collector
detecting an init failure.

The daemon could still crash because of compiler or run-time
system bugs, or errors in the use of the Foreign Function
Interface, but this is much less likely and this CL does
not make it any worse.

BUG=none
TEST=manual

Change-Id: I7ff0c8725cd1e69f2b6a9c1bf97528b4cae2476e
Reviewed-on: https://chromium-review.googlesource.com/c/1294969
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>
Commit-Queue: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>