init: fix detection of rootdev in FW updaters
The firmware updater scripts can optionally force a reboot right after
the update, except when the device is booting from a removable device.
Fix the detection of the device that we're booting from.
Also, when the script detects that it's booting from a removable device,
log the devices in question.

BUG=b:119254302
TEST=works as expected on device with rootfs verification OFF

Change-Id: I94e781fc81a086d0cfb86b8382d8b9f5cb7b5295
Reviewed-on: https://chromium-review.googlesource.com/c/1327664
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>
Commit-Queue: Nicolas Norvez <norvez@chromium.org>
Tested-by: Nicolas Norvez <norvez@chromium.org>