vm_tools: concierge: Properly clean up child exits
If crosvm exits on its own (for example, due to a guest kernel panic)
then concierge doesn't delete the TerminaVm pointer correspending to
that VM from the `termina_vms_` field in the Service class.  This is a
problem because the pointer is now pointing to freed memory and will
cause a segfault if we try to dereference it later.

Make sure to delete the TerminaVm pointer and also notify cicerone that
the VM was stopped.

BUG=chromium:946852
TEST=stop a running vm with maitred_client, run
     `concierge_client --sync_times`, and observe that concierge does
     not crash

Change-Id: I75f67f1e689a63d0be074962bb986b5ae688140d
Reviewed-on: https://chromium-review.googlesource.com/1541849
Commit-Ready: Chirantan Ekbote <chirantan@chromium.org>
Tested-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>
(cherry picked from commit f2be8f36a509f52846aa23bfd91acdacea56389c)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1548771
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Commit-Queue: Chirantan Ekbote <chirantan@chromium.org>