shill: point ethernet_any to first available Ethernet service
There are a set of assumptions that Chrome makes when applying
service policies that aren't currently upheld with regards to the
ethernet_any service (i.e. ethernet_any is not in the
ServiceCompleteList, and the service path returned when
ethernet_any is configured is not the same one returned after a
device is plugged in). We address that in this change by pointing
ethernet_any to the first available Ethernet service (dictated by
the service sorting heuristic).

The algorithm is as follows: when a new Ethernet service is created
first check to see if there are currently any Ethernet services. If
there aren't, simply assign the new Ethernet device to the existing
ethernet_any service instead of instantiating a new one. This way,
the service path remains the same if ConfigureServiceForProfile was
called before a device was plugged in. If this device is removed at
some point, point ethernet_any at the first Ethernet service in the
list. This service will now act as the ethernet_any service. If, at
some point, the last Ethernet device is unplugged, instead of
destroying the service, convert it back to the generic ethernet_any
service by removing the associated device and setting the storage
identifier to "ethernet_any".

TODO: if there are multiple Ethernet services and the default service
changes, the GUID remains the same. Chrome may expect the GUID and
service path to be consistent, so this will probably need to be
corrected.

BUG=chromium:949553,chromium:950067,chromium:988599
TEST=1) unittests
     2) deployed to DUT, configured static IP, and did the following:
         a) plugged in eth0
         b) plugged in eth1
         c) removed eth0
         d) removed eth1
        Confirmed that ethernet_any pointed to the correct
        service at all steps and that static IP transferred between
        the two Ethernet services when it should.
     3) ConfigureServiceForProfile still works and now returns
        the same service path.
     4) network_VPNConnect no longer persists 10.9.8.2 IP

Change-Id: Ib2f3c398cc188db2f307df68aeadbd80e9f65d98
Reviewed-on: https://chromium-review.googlesource.com/1580541
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
(cherry picked from commit ab614c8a72b5d1f4aed4b3dd9833c2145c239e0d)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1763916
Commit-Queue: Matthew Wang <matthewmwang@chromium.org>