init: chromeos_startup: Avoid clobber when TPM ownership is unknown
To determine if TPM is owned, we try to read a file and that ends up
sending a tpmc command to TPM. In case the command fails to execute,
the function is_tpm_owned was returning "false", which would trigger
the device clobber if install attributes is present.

Here we change this, to consider that TPM is owned if the command fails,
thus avoiding the device clobber in such case. The only risk this has is
the case when TPM is cleared, install attributes is present and TPM
command to determine if owned fails. In that case we end up with some
dead data on disk. This is much less problematic than having the device
clobber by mistake.

BUG=chromium:929695
TEST=None

Change-Id: I17598c560c115cb6fd2d2c6edce4b14f88ad9c37
Reviewed-on: https://chromium-review.googlesource.com/1626095
Commit-Ready: Igor <igorcov@chromium.org>
Tested-by: Igor <igorcov@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Igor <igorcov@chromium.org>
(cherry picked from commit 066508b6826c2266f1a49dc3fa649f5121b837bc)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1667002
Commit-Queue: Igor <igorcov@chromium.org>