power: add keyboard rule for convertible that disables volume buttons.
We need another internal keyboard variant that uses the default inhibit
behavior for the internal keyboard when it is not usable. If we leave
the Chrome to inhibit the keyboard, then the volume up, volume down, and
power button key presses will be allowed even during tablet mode. If the
volume buttons are not separate side buttons, then it is weird for every
key to be disabled but the volume key.

BRANCH=R76
BUG=b:136110848
TEST=with this new udev rule instead of the side-button udev rule, the
volume up and down buttons do not work in tablet mode (which is the
correct behavior).

Change-Id: I5f05e062b69a04c61b9ad45f4f4451f3eef0957a
Signed-off-by: Jett Rink <jettrink@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1680908
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Raul E Rangel <rrangel@chromium.org>
(cherry picked from commit 3464b6510cc42c0ae92eee82aa51e398f07a6b53)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1687599