u2fd: Always enable user keys
This change deprecates the "legacy" code path,
and enables legacy fallback if users have not
swiched explicitly to the new code path.

This is a forced migration to the new code path,
while still supporting legacy key handles for
authentication.

For a user who has previously enabled U2F with
'u2f' or 'g2f' mode (but not switched to user keys),
this change will cause new registrations to be
user-specific, but old keys can still be used to
authenticate. No user action is required for this
behavior.

BUG=b:141269706
TEST=flash u2fd prior to this change
     set flags to --u2f (without --user_keys)
     generate legacy key handle
     flash u2fd with this change
     check legacy kh still works
     check new kh is user-specific
     set flags to --u2f --user_keys
     check legacy kh does not work

Change-Id: I3d273e5233427d8641532e3dec9664107a32672d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1824399
Reviewed-by: Wei-Cheng Xiao <garryxiao@chromium.org>
Commit-Queue: Louis Collard <louiscollard@chromium.org>
Tested-by: Louis Collard <louiscollard@chromium.org>