crash: Filter vm crashes to exclude containers.
We add a coarse filter that checks if the crashed app has a PID
namespace outside the root namespace. If it does this implies it runs in
a container and we do not report its crashes.

This is a coarse filter which we will use for now to avoid noise, in
subsequent CLs we will refine the filter to allow some container-based
crashes to be reported (e.g. garcon, sommelier).

BUG=chromium:703926
TEST=Manually tested

Change-Id: I61179a78294d5cd0bafb8de7d33f52fd13bc4c3a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2024207
Tested-by: Fergus Dall <sidereal@google.com>
Commit-Queue: Fergus Dall <sidereal@google.com>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2036864
Reviewed-by: Fergus Dall <sidereal@google.com>