shill: send PMK as byte array
We would like to continue to support sending raw PMKs (64 character hex
string) through shill to supplicant. We merged crrev.com/c/323775 to
wpa_supplicant 2.6 to support this but it was dropped because the only
use case was fast reconnect (crbug.com/955832).

Through discussions in $BUG, we decided to revive this functionality and
realized that wpa_supplicant already supports this if the PMK is sent as
a byte array instead of a string. This change checks to see if the PSK
is a raw PMK, sends it as a byte array if it is, and sends it as a
normal string if it isn't.

BUG=chromium:1041057
TEST=1) unit tests
     2) network_WiFi_SimpleConnect.wifi_checkRawPMK
     3) manually enter raw PMK

Change-Id: I935b870fffd845ff1bbde66526dc57549ba099ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2031499
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
(cherry picked from commit 0635ae6ff668229112f0cb92a129e9b1d8d76b47)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2050389
Commit-Queue: Brian Norris <briannorris@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>