crash: Give crash_sender CAP_FOWNER
crash_sender needs to be able to delete files in the daemon-store
crash spool directory. For the time being crash_sender only runs as
root, which means this directory must either be owner by root or
crash_sender needs to get CAP_FOWNER, which allows for overriding that
check.

After discussing with jorgelo@, we've decided that the second is the
preferred solution, as it makes it more explicit what the problem is
when it comes time to fix it, and doesn't change any persistent state.

BUG=chromium:1067192
TEST=Create VM crash, trigger upload, check that it gets deleted

Change-Id: Ic06bec8c5c3f12db676087cbebe6c5821f081fa5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2134040
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Tested-by: Fergus Dall <sidereal@google.com>
Commit-Queue: Fergus Dall <sidereal@google.com>
(cherry picked from commit 125dc9a3b7dfc73c3942e1878a766e7800283b1f)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2147973
Reviewed-by: Fergus Dall <sidereal@google.com>