permission_broker: separate lifeline_fd for usb driver tracking.
The current permission_broker detaches existing usb drivers
from device interfaces when opening a USBDEVFS path for the
client, with the intention of reattaching those drivers when
the client closes the path. This mechanism doesn't seem to
have worked as intended. Prior to M85 the failure was
relatively harmless - clients closing the device path would
never be noticed, and the drivers would never be reattached
until the device was physically unplugged/replugged.

With M85 (or M86) we started to see device drivers
re-attaching as soon as a device path was shared with a VM.
This meant that the VM was unable to attach its own drivers
to an interface using ioctl USBDEVFS_CONNECT because the
interfaces were already reclaimed by the host kernel drivers.

This CL returns us to the status quo ante of no kernel drivers
re-attaching while providing a new ClaimDevicePath dbus method
with a separate lifeline fd for monitoring the lifetime of the
client's intended claim on the device. FileDescriptorWatcher
WatchReadable is used to see when the source of the pipe is
closed.

BUG=chromium:1128752
TEST=Manual testing

Change-Id: I043289bce724c2ee832174271de2b37f6c842296
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2437829
Tested-by: Nicholas Verne <nverne@chromium.org>
Commit-Queue: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Auto-Submit: Nicholas Verne <nverne@chromium.org>
(cherry picked from commit 2e60b0ded2d0959f19a4f4468ee4ef0b535b66db)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2501088
Reviewed-by: Nicholas Verne <nverne@chromium.org>
Commit-Queue: Nicholas Verne <nverne@chromium.org>