u2fd: Query cryptohomed and biod for authenticator availability
The previous implementation of IsUvpaa
(IsUserVerifyingPlatformAuthenticatorAvailable) doesn't work with
Chrome because the IsUvpaa call in Chrome is synchronous and the
previous implementation relies on asynchronous response from Ash.

This change re-implements the IsUvpaa handler to talk to crytohomed
for PIN and biod for fingerprint instead, so that it can be
synchronized.

Ideally in the future we would have a refactor in Chrome to allow
for asynchronous IsUvpaa.

Also refactor GetWebAuthnSecret to use the simpler
org::chromium::CryptohomeInterfaceProxy.

BUG=b:144861739
TEST=Using webauthntool --is_uvpaa, verified that IsUvpaa returned
     true if either PIN or fingerprint is set up, false otherwise.
TEST=With a Chrome change that calls IsUvpaa, verify that the developer
     console command
     "PublicKeyCredential.isUserVerifyingPlatformAuthenticatorAvailable()
          .then(console.log)"
     gives true only if either fingerprint or PIN is set up.
TEST=With a Chrome change that calls IsUvpaa, verify that WebAuthn
     requests are only dispatched if either fingerprint or PIN is set
     up.
TEST=Try MakeCredential flow and succeed, thus proving GetWebAuthnSecret
     functionality isn't changed.

Change-Id: I8573eb2d4f193374332b11ae6e40a8299ce740e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2548560
Reviewed-by: Yicheng Li <yichengli@chromium.org>
Reviewed-by: John L Chen <zuan@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Tested-by: Yicheng Li <yichengli@chromium.org>
Commit-Queue: Yicheng Li <yichengli@chromium.org>
(cherry picked from commit 67abd18dc2839db866a0aa86dea3d4edd75aacf1)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2562689