typecd: Add current mode field for Port
The port should keep track of which mode it's currently in so that
PortManager can avoid calling "EnterMode" on a port which has already
entered a mode. Add a field (with getter and setter) to store the
current Type C mode in the Port.

Also update the TypeCMode enum to list a "NONE" mode which is what a
port is initialized to.

BUG=b:171725237
TEST=Unit tests should still pass.

Change-Id: If8bb1a9a2f36996ec562a8c5832126735f787356
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2626070
Tested-by: Prashant Malani <pmalani@chromium.org>
Reviewed-by: Prashant Malani <pmalani@chromium.org>
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Commit-Queue: Prashant Malani <pmalani@chromium.org>
(cherry picked from commit e269659f766c862113dd7cf14e8485de161487d6)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2831336
Tested-by: Cheng Yueh <cyueh@chromium.org>
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Commit-Queue: Cheng Yueh <cyueh@chromium.org>