[LTS-M86] cryptohome: stop acl patching for chaps
The recursive ACL patching was added as a presonse to an unknown and
un-debugged issue where the files under the chaps directory ended with
wrong permissions ([1]). That is a unfavourable pattern to have in the
code base as it just masks problems not fixing. On top of that, the
recursive permission application is known to be an enormous source of
vulnerabilities. Keeping that in mind, the CL removes the recursive
permission patching. If the issue which triggered addition of the
behaviour is still present in the code base, we shall investigate it
properly instead of making cryptohome attend to things which is not even
its business.

[1] https://chromium-review.googlesource.com/c/chromiumos/platform/cryptohome/+/169750/

BUG=chromium:1168541
TEST=unittest

Change-Id: I6d2855ff85dfe53ca57f6e9b23f19a7d75eac304
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2643854
Tested-by: Daniil Lunev <dlunev@chromium.org>
Commit-Queue: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Greg Kerr <kerrnel@chromium.org>
(cherry picked from commit 60e49cf2138634002c1c7d6a0aacb67658d118d7)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2752802
Commit-Queue: Victor-Gabriel Savu <vsavu@google.com>
Tested-by: Victor-Gabriel Savu <vsavu@google.com>
Reviewed-by: Victor-Gabriel Savu <vsavu@google.com>