u2fd: Disable webauthn handler in u2f/g2f mode (for M89)
We are in the late stages of developing and testing WebAuthn support
in U2F/G2F mode. We were unable to finish this support in M89, so
we want M89's behavior to be the same as M88, which is to disable
WebAuthnHandler in U2F/G2F mode. Unfortunately, one change,
https://crrev/c/2598176 went in M89. Normally we should revert that
change on ToT, then merge the revert to M89, but ToT has diverged
from M89 so much that it's misleading and error-prone to revert that
change.

So we just create this new change to achieve the same thing. This
change is intended to be submitted on ToT, tested, merged to M89,
and reverted on ToT.

BUG=b:179294931
TEST=Enter u2f mode.
     Platform authenticator doesn't show up when doing MakeCredential
     with attachment=platform or attachment=undefined.

Change-Id: I4d22d0e8fd19250a156cf1d04e8be4946481178f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2673167
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Tested-by: Yicheng Li <yichengli@chromium.org>
Commit-Queue: Yicheng Li <yichengli@chromium.org>
(cherry picked from commit 5aa5deccaddb3fea4df7aeae9ce54f8092b06ae1)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2682849
Reviewed-by: Yicheng Li <yichengli@chromium.org>