[LTS-M86] cryptohome: skip creating daemon-store source if existing
An unexpected side effect of the recent change to use safe directory
creation [1] is that now there is a more rigorous check on whether the
acl's of the directories which shall be present match the expected ones.
The older code didn't check for the permissions match, but
SafeFd::MakeDir does so. This may cause an inability to login if
something changes the ACLs in the meantime. This seems to be happening
to one of the daemons (see the bug and tactical fix [2]). This change
makes us skip the already existing directories not to impose
side-effects which were not present in the previous code.

[1] https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2643664
[2] http://shortn/_NEWoQPWELu

BUG=chromium:1176379
TEST=unittest

Change-Id: Ifca2bd9db50f1bbe23055b9595f310072570f3b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2686855
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Daniil Lunev <dlunev@chromium.org>
Tested-by: Daniil Lunev <dlunev@chromium.org>
(cherry picked from commit b98321b6f74e277b46a477ea260a42c9d23a8662)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2757250
Commit-Queue: Victor-Gabriel Savu <vsavu@google.com>
Reviewed-by: Victor-Gabriel Savu <vsavu@google.com>
Tested-by: Victor-Gabriel Savu <vsavu@google.com>