shill: manager: remove a false error.
When querying a service using its storage identifier, the Manager was
replying with an error when not found. As it is not really an error but
something common, the trace and the error are now removed. The CL also
fixes a small crash on unchecked NULL value that appeared when fixing
the trace.

BUG=b:178697621
TEST=FEATURES="test" emerge-${BOARD} shill
TEST=Check with D-Bus calls that setting the always-on VPN mode and
  service is working correctly.

Change-Id: I673852d6ed80a038b2c16dcf0c08f012eb44eed4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2696330
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Tested-by: Damien Dejean <damiendejean@google.com>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
(cherry picked from commit b37510c002dae8c963345734ccbe673ddd861187)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2739623
Commit-Queue: Damien Dejean <damiendejean@google.com>