modemfwd: fix (again) modem SIM swap detection
Since the SIM handling has changed on the Shill side for multi-SIM
support, when we are receiving the notification for the SIM ICCID
change, the HomeProvider property might not contain the new carrier UUID
yet. As a consequence, we cannot select the proper modem carrier
customization.

In order to close the race condition, listen to the cellular Device
HomeProperty rather than the ICCID property.

BUG=b:181644166
TEST=manual, hotswap the SIM in the L850 modem and look carefully at the
/var/log/net.log file along with the modemfwd traces.

Change-Id: I6f56d09fc1807afa0524dc8cbe99d7cfaf12192d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2729379
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2910413
Tested-by: Stimim Chen <stimim@chromium.org>
Commit-Queue: Stimim Chen <stimim@chromium.org>