modemfwd: decouple modem detection from the flashing process
Create a new event 'OnModemDeviceSeen' when ModemManager has seen the
device on the bus and shill has created the corresponding 'Device' object.
From this point, we can mark the modem as 'seen' and no longer need to
try rebooting/recovering the device.
Keep the previous event (ie 'OnModemAppeared') to do the actual version
detection and flashing as it pops *after* we actually know the carrier
ID to decide about the needed firmware, but rename it as
'OnModemCarrierIdReady'.

This change should solve the case where the machine is booted without
SIM, so we never get the modem carrier ID but still don't need to
recover the modem as it has been seen properly.

BUG=b:182833475
TEST=On Jinlon, boot without SIM inserted and no longer see spurious
modem reboot in the logs.
TEST=On Jinlon, verify that the firmware re-flashing still happens when
swapping SIM.

Change-Id: Ia1c3d1ccb413bccd28a1a7e5900f8a88def5c36b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2767229
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3066815
Tested-by: Stimim Chen <stimim@chromium.org>
Commit-Queue: Stimim Chen <stimim@chromium.org>