modemfwd: fix journal playback for compressed firmwares
The journal code was broken for compressed firmwares.
It was not trying to uncompress them before sending them to the flashing
helper.

BUG=None
TEST=manual, on Boten, run in crosh 'modem force-flash usb:2cb7:0007'
interrupt it with 'kill -9' of modemfwd. Re-start modemfwd and see the
journal traces and a successful re-flashing.

Change-Id: I67e37485bdf1a59a1c4f4d2b5c6af0acb02c88bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2794441
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3068660
Commit-Queue: Stimim Chen <stimim@chromium.org>
Tested-by: Stimim Chen <stimim@chromium.org>