modemfwd: fix second invocation of 'modem force-flash'
When running twice 'modem force-flash' in crosh, the second call was not
triggering the actual modem firmware flashing.
This was actually due to a 'trap' in the base::UnguessableToken API:
the default constructor UnguessableToken() always return zeros.
The .Create() method needs to be called to get an UnguessableToken which
is actually random.
As a consequence the 'equipment_id_' was a constant (00...00) in the
StubModem used by ForceFlash code and on the second call,
ShouldFlashMainFirmware() would be false.

BUG=b:184139554
TEST=on Boten with L850-gl modem, run twice in crosh:
'modem force-flash usb:2cb7:0007', inspect the syslog and see the modem
was flashed and rebooted twice.

Change-Id: Iffd28d7c2b092337fbf13c7791dc67f2e305369e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2797337
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3068662
Commit-Queue: Stimim Chen <stimim@chromium.org>
Tested-by: Stimim Chen <stimim@chromium.org>