shill: rollback the Device ID for USB
In the device ID, the 'location' of the device was appended recently to
the exported string, e.g. for USB devices 'usb:_vid_:_pid_ (Unknown location)'
instead of 'usb:_vid_:_pid_'.
Revert to the previous string for USB (and all devices with 'Unknown
location') but keep the location when it is known (namely for PCI
devices).

For cellular, this device ID is sent over D-Bus in the Cellular.DeviceID
property, and the modemfwd daemon is using it to match the proper
modem firmware blobs and modem update helpers using the platform
specific manifests containing those device_id for each supported modem.
Add a unit-test to ensure this string identifier is not changing in the
future.

BUG=b:184943818
TEST=unit-tests
TEST=manual, on Metaknight, boot without SIM card and verify the logs,
there is no spurious modem reboot and the device id looks ok in the
modemfwd logs.

Change-Id: Ib244007aea0078a00052f88b64a21045c494fe12
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2817768
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>
(cherry picked from commit e3131e8e8455e15204e65e38549e566b11f59cdc)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2830931
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>