shill: vpn: save always-on config correctly.
The profile is usually saved on Shill shutdown. It is not enough for
always-on VPN as a user may set a configuration through the UI and
logout. In such case, the profile is removed but not saved. this CL
ensure the profile is saved after a valid always-on VPN configuration is
set.

BUG=b:188776416
TEST=shill unit tests;
  - Set a VPN client as always on, logout, login and check it connects
  automatically
  - Set a VPN client as always on, reboot. After login ensure the
  clients is connected. Set another client as always-on VPN, logout and
  login back. Ensure the last configure client is connected.

Change-Id: I4cc0d9d6f98f4b57f8bf092b8ea6e3ec8000f76f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2912101
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Tested-by: Damien Dejean <damiendejean@google.com>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
(cherry picked from commit 9baa29c8f4330a316f37703d96fad9859bfa9c5f)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2919687
Commit-Queue: Damien Dejean <damiendejean@google.com>