arc: Fix arcvm-boot-notification-server crash
This change modifies arcvm-boot-notification-server so that, when it is
sending props to the guest, it will continue accepting connections until
it is able to write the props successfully.

This is to address a race between the previous and current VM that
happens when Chrome restarts while the first VM is in its early stages
of boot up. Chrome will not kill the previous VM when shutting down, so
it will continue booting until arcvm-boot-notification-client connect().
However, by the time the previous VM reaches connect(), a new
arcvm-boot-notification-server has started. arcvm-boot-notification-server
only expects one connect request from the guest, but it will fail
because the previous VM is killed before the server can accept(). In
order to allow the current VM to connect and boot successfully, we need to
accept() again.

BUG=b:188450841
TEST=To reproduce crash, apply gpaste/5685150860967936 patch to Chromium.
 Apply this change to arcvm-boot-notification-server, and verify that it
 does not crash

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2939009
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Cherie Cheung <cherieccy@google.com>
Tested-by: Willie Koomson <wvk@google.com>
Commit-Queue: Yusuke Sato <yusukes@chromium.org>
(cherry-picked from commit d5a5175a0dcff0908af7d7da9d9cc64677584337)

Change-Id: I3a2ae55766914efe3480987cb10164014be4e87b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2960516
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Commit-Queue: Willie Koomson <wvk@google.com>
Tested-by: Willie Koomson <wvk@google.com>