modemfwd: skip clearing attach APN when carrier UUID is unknown
When running the 'l850gl-helper' with an carrier UUID, we are getting an
error, e.g.
Failed to perform "clear_attach_apn=" on the modem
Clear attach APN failed for current carrier.

As we are actually the attach APN only for a set of known carrier UUIDs,
we can skip the call in this case.

BUG=b:191990088
TEST=On Jinlon, run 'modem force-flash usb:2cb7:0007' and no longer see
'Clear attach APN failed for current carrier' error.

Change-Id: I5e886ea931c11d226bdfd129f9eb24e7b13cfd3e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2991632
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3068666
Commit-Queue: Stimim Chen <stimim@chromium.org>
Tested-by: Stimim Chen <stimim@chromium.org>