dns-proxy: Fix shill client usage
In short, the shill client does not need to be cleared and reset when/if
the service itself is detected to be lost. Rather, we just need to keep
track of it to avoid using it in these cases. So this patch ensures that
the OnAvailableCallback is hooked if shill is lost and initializes the
client just once on setup.

BUG=b:193712306
BUG=b:193090160
BUG=b:194124278
TEST=units
TEST=revert seccomp patch, with new shill client patch, cannot repro
crash

Change-Id: I260681157952905ab26aeccb384fa8718d551e9c
Previous-Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3037765
(cherry picked from commit 6481d9942587ed7191c54364781bad394018e141)
Previous-Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3183500
Previous-Reviewed-on: https://chromium-review.googlesource.com/3183500
(cherry picked from commit bc5b3b6ec89afe48ede92ed2095c4a378131ebc8)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3187771
Tested-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-by: Garrick Evans <garrick@chromium.org>
Commit-Queue: Sarthak Kukreti <sarthakkukreti@chromium.org>