dns-proxy: Fix flaky reply in c-ares client.
This issue was first observed on a dood/octopus board where DNS
resolution via the system proxy was broken, until, it appeared, the user
logged in. In particular, calls to dig would return these
warnings:

;; Warning: ID mismatch: expected ID 12788, got 0
;; Warning: query response not set

Both indicating cleared bits in the answer message. This problem was not
initially observable on other boards like rammus or trogdor but I was
able eventually to get it to show up on a trogdor. Anyway, as explained
in the bug, the root cause is the passing of the original raw pointer
that c-ares delivers to the send callback, through a PostTask to the
HandleResult closure. Dumping the buffers revealed the message was
corrupted it the handler and subsequently returned to the caller. This
patch also switched the BindRepeating used in PostTask to BindOnce.

This is also the root cause of the issue first observed in b/193752871

Incidentally, the reason DNS worked after logging in was that DoH was
enabled and the queries were flowing through the CURL client.

BUG=b:194022247
BUG=b:193752871
BUG=b:194124278
TEST=deployed to both trodgor and octopus boards, verified message
buffer in posted handler, and dig receives the proper answer.

Change-Id: I56f44c7c33844ec91edb498b9e63875a59c6c1e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3038129
Tested-by: Garrick Evans <garrick@chromium.org>
Commit-Queue: Garrick Evans <garrick@chromium.org>
Reviewed-by: Jason Jeremy Iman <jasongustaman@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
(cherry picked from commit 9a5904792d728bf351851645439da4bb78c9db61)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3183501
Commit-Queue: Sarthak Kukreti <sarthakkukreti@chromium.org>
Tested-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-by: Garrick Evans <garrick@chromium.org>